change KODI/VDR startup (see https://www.vdr-portal.de/forum/thread/135091-installation-eines-vdr-plugins-nativ-auf-coreelec-boxen/?postID=1393949#post1393949)

+++ packages/vdr/_config/bin/switch_kodi_vdr.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 
 #
-# This script will be started after the file /storage/.cache/switch_kodi_vdr has been changed
+# This script will be started by systemd (switch_kodi_vdr.path via switch_kodi_vdr.service) after the file /storage/.cache/switch_kodi_vdr has been changed
 #
 
 # Read .profile
@@ -10,18 +10,28 @@
 # Start either Kodi or VDR on request
 . /storage/.cache/switch_kodi_vdr
 
+
 if [ "${START_PRG}" = "vdr" ]; then
-  systemctl stop kodi
-  if [ ! -z ${SWITCH_VDR_SCRIPT} ]; then
-    eval ${SWITCH_VDR_SCRIPT} attach
-  else
-    systemctl start vdropt
-  fi
+   #check if Kodi is already running (should never be the case after reboot)
+   if pgrep "kodi.bin" > /dev/null ; then
+      systemctl stop kodi
+      /storage/.config/vdropt/ClearOSD.sh || true
+   fi
+   if [ ! -z ${SWITCH_VDR_SCRIPT} ] && pgrep "vdr" > /dev/null ; then
+      eval ${SWITCH_VDR_SCRIPT} attach
+   else
+      systemctl start vdropt
+   fi
 elif [ "${START_PRG}" = "kodi" ]; then
-  if [ ! -z ${SWITCH_VDR_SCRIPT} ]; then
-    eval ${SWITCH_VDR_SCRIPT} detach
-  else
-    systemctl stop vdropt
-  fi
-  systemctl start kodi
+   #check if vdr is running
+   if pgrep "vdr" > /dev/null ; then
+      if [ ! -z ${SWITCH_VDR_SCRIPT} ]; then
+         eval ${SWITCH_VDR_SCRIPT} detach
+         /storage/.config/vdropt/ClearOSD.sh || true
+      else
+         systemctl stop vdropt
+      fi
+   fi
+   systemctl unmask kodi
+   systemctl start kodi
 fi
